+++ example/src/index.js
@@ -12,6 +12,7 @@ const client = new TOS({
   },
   requestTimeout: -1,
   bucket,
+  enableCRC: true,
 });
 
 const listBucketsDom = document.querySelector('#list-buckets');