+++ CHANGELOG.md
@@ -2,6 +2,18 @@
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+Unreleased
+
+### Added
+
+- disable request redirect for 3xx response status code
+- support crc64
+- add `(get|put)BucketVersioning` methods to modify bucket versioning
+
+### Fixed
+
+- set user-agent request header
+
 ## [2.1.13] - 2022-10-30
 
 ### Added